feat: add clipboard for issue lists
