+++ public/src/app/issue-list/issue-list.component.html
@@ -8,4 +8,5 @@
       </span>
     </li>
   </ul>
+  <app-clipboard [text]="getIssueListText()" title="Copy list of issues to clipboard"></app-clipboard>
 </div>
